moved helper functions to their own files
