+++ keystrokes/src/utils/utils.js
@@ -1,0 +1,6 @@
+export function pickRandomElement(givenThis){
+    const randomKey = givenThis.state.keys[
+        Math.floor(Math.random() * givenThis.state.keys.length)
+    ];
+    givenThis.setState({ key: randomKey });
+};
